src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
pymupdf==1.24.3
pymongo==4.8.0
motor==3.4.0
openai==2.0.1
cohere==5.5.8
qdrant-client==1.10.1
SQLAlchemy==2.0.36
asyncpg==0.30.0 
alembic==1.14.0
psycopg2==2.9.10
prometheus-client==0.19.0
starlette-exporter==0.17.1
fastapi-health==0.4.0
celery==5.5.3
redis==6.2.0
kombu==5.5.4
billiard==4.2.1
vine==5.1.0
flower==2.0.1